feat: Rework Application Launcher Admin UI - MEED-8633 - Meeds-io/MIPs#185 This change will change the App Launcher Admin UI.
