+++ app-center-services/src/main/resources/locale/addon/appcenter_ar.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=إعدادات
 appCenter.adminSetupForm.maxFavoriteApps=عدد الطبيقات المفضلة المسموح به
 appCenter.adminSetupForm.defaultAppImage=أيقونة التطبيق الأساسية
 appCenter.adminSetupForm.applications=التطبيقات
-appCenter.adminSetupForm.addNewApp=إضافة تطبيق
 appCenter.adminSetupForm.editApp=تعديل التطبيق
 appCenter.adminSetupForm.modal.DeleteApp=هل أنت متأكد؟
 appCenter.adminSetupForm.createNewApp=إضافة تطبيق
