+++ app-center-services/src/main/resources/locale/addon/appcenter_ca.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Configuració
 appCenter.adminSetupForm.maxFavoriteApps=Nombre de favorits permesos
 appCenter.adminSetupForm.defaultAppImage=Icona de l'aplicació per defecte
 appCenter.adminSetupForm.applications=Aplicacions
-appCenter.adminSetupForm.addNewApp=Afegeix una aplicació
 appCenter.adminSetupForm.editApp=Edita l'aplicació
 appCenter.adminSetupForm.modal.DeleteApp=Estàs segur ?
 appCenter.adminSetupForm.createNewApp=Afegeix una aplicació
