+++ app-center-services/src/main/resources/locale/addon/appcenter_co.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Settings
 appCenter.adminSetupForm.maxFavoriteApps=U numeru di preferiti permessi
 appCenter.adminSetupForm.defaultAppImage=Icona di l'applicazione predeterminata
 appCenter.adminSetupForm.applications=Applicazioni
-appCenter.adminSetupForm.addNewApp=Aghjunghjite una applicazione
 appCenter.adminSetupForm.editApp=Edite l'applicazione
 appCenter.adminSetupForm.modal.DeleteApp=Sì sicuru ?
 appCenter.adminSetupForm.createNewApp=Aghjunghjite l'applicazione
