+++ app-center-services/src/main/resources/locale/addon/appcenter_cs.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Nastavení
 appCenter.adminSetupForm.maxFavoriteApps=Počet povolených oblíbených
 appCenter.adminSetupForm.defaultAppImage=Výchozí ikona aplikace
 appCenter.adminSetupForm.applications=Aplikace
-appCenter.adminSetupForm.addNewApp=Přidat aplikaci
 appCenter.adminSetupForm.editApp=Upravit aplikaci
 appCenter.adminSetupForm.modal.DeleteApp=Jste si jisti?
 appCenter.adminSetupForm.createNewApp=Přidat aplikaci
