+++ app-center-services/src/main/resources/locale/addon/appcenter_de.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Einstellungen
 appCenter.adminSetupForm.maxFavoriteApps=Anzahl erlaubter Favoriten
 appCenter.adminSetupForm.defaultAppImage=Standardanwendungs-Symbol
 appCenter.adminSetupForm.applications=Apps
-appCenter.adminSetupForm.addNewApp=App hinzufügen
 appCenter.adminSetupForm.editApp=App bearbeiten
 appCenter.adminSetupForm.modal.DeleteApp=Sind Sie sicher?
 appCenter.adminSetupForm.createNewApp=App hinzufügen
