+++ app-center-services/src/main/resources/locale/addon/appcenter_el.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Ρυθμίσεις
 appCenter.adminSetupForm.maxFavoriteApps=Αριθμός επιτρεπόμενων αγαπημένων
 appCenter.adminSetupForm.defaultAppImage=Προεπιλεγμένο εικονίδιο εφαρμογής
 appCenter.adminSetupForm.applications=Εφαρμογές
-appCenter.adminSetupForm.addNewApp=Προσθήκη εφαρμογής
 appCenter.adminSetupForm.editApp=Επεξεργασία εφαρμογής
 appCenter.adminSetupForm.modal.DeleteApp=Είστε σίγουροι?
 appCenter.adminSetupForm.createNewApp=Προσθήκη εφαρμογής
