+++ app-center-services/src/main/resources/locale/addon/appcenter_en.properties
@@ -37,7 +37,7 @@ appCenter.adminSetupForm.generalSettings=Settings
 appCenter.adminSetupForm.maxFavoriteApps=Number of favorites allowed
 appCenter.adminSetupForm.defaultAppImage=Default application icon
 appCenter.adminSetupForm.applications=Applications
-appCenter.adminSetupForm.addNewApp=Add an application
+appCenter.adminSetupForm.add=Add
 appCenter.adminSetupForm.editApp=Edit application
 appCenter.adminSetupForm.modal.DeleteApp=Are you sure ?
 appCenter.adminSetupForm.createNewApp=Add application
@@ -59,6 +59,7 @@ appCenter.adminSetupForm.descriptionPlaceHolder=Describe the purpose of the appl
 appCenter.adminSetupForm.mandatory=Mandatory
 appCenter.adminSetupForm.active=Active
 appCenter.adminSetupForm.mobile=Mobile
+appCenter.adminSetupForm.loadMore=Load more
 appCenter.adminSetupForm.table.switch.mobile.tooltip=Visibility on mobile
 appCenter.adminSetupForm.table.editApp.tooltip=Edit application
 appCenter.adminSetupForm.table.deleteApp.tooltip=Delete application
@@ -71,11 +72,13 @@ appCenter.adminSetupForm.save=Save
 appCenter.adminSetupForm.cancel=Cancel
 appCenter.adminSetupForm.edit=Edit
 appCenter.adminSetupForm.modal.delete=Delete
+appCenter.adminSetupForm.modal.system.delete=This product application cannot be deleted
 appCenter.adminSetupForm.browse=Browse
 appCenter.adminSetupForm.error=Server error
 appCenter.adminSetupForm.unauthorized=An application with the same name or URL already exists
 appCenter.adminSetupForm.modal.confirmDelete=Do you really want to delete this application? This process can not be undone.
 appCenter.adminSetupForm.noApp=No application
+appCenter.adminSetupList.name=Name
 appCenter.adminSetupList.application=Application
 appCenter.adminSetupList.avatar=Icon
 appCenter.adminSetupList.actions=Actions
