+++ app-center-services/src/main/resources/locale/addon/appcenter_es_ES.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Ajustes
 appCenter.adminSetupForm.maxFavoriteApps=Número de favoritos permitidos
 appCenter.adminSetupForm.defaultAppImage=Icono de aplicación por defecto
 appCenter.adminSetupForm.applications=Aplicaciones
-appCenter.adminSetupForm.addNewApp=Añadir una aplicación
 appCenter.adminSetupForm.editApp=Editar aplicación
 appCenter.adminSetupForm.modal.DeleteApp=¿Estás seguro?
 appCenter.adminSetupForm.createNewApp=Añadir aplicación
