+++ app-center-services/src/main/resources/locale/addon/appcenter_et.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Seaded
 appCenter.adminSetupForm.maxFavoriteApps=Lubatud lemmikute arv
 appCenter.adminSetupForm.defaultAppImage=Vaikimisi rakenduse ikoon
 appCenter.adminSetupForm.applications=Rakendused
-appCenter.adminSetupForm.addNewApp=Lisage rakendus
 appCenter.adminSetupForm.editApp=Redigeeri rakendust
 appCenter.adminSetupForm.modal.DeleteApp=Oled sa kindel ?
 appCenter.adminSetupForm.createNewApp=Lisa rakendus
