+++ app-center-services/src/main/resources/locale/addon/appcenter_fa.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=تنظیمات
 appCenter.adminSetupForm.maxFavoriteApps=تعداد موارد دلخواه مجاز است
 appCenter.adminSetupForm.defaultAppImage=نماد برنامه پیش فرض
 appCenter.adminSetupForm.applications=برنامه های کاربردی
-appCenter.adminSetupForm.addNewApp=یک برنامه کاربردی اضافه کنید
 appCenter.adminSetupForm.editApp=ویرایش برنامه
 appCenter.adminSetupForm.modal.DeleteApp=مطمئنی ?
 appCenter.adminSetupForm.createNewApp=افزودن برنامه
