+++ app-center-services/src/main/resources/locale/addon/appcenter_fi.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Asetukset
 appCenter.adminSetupForm.maxFavoriteApps=Suosikkien määrä sallittu
 appCenter.adminSetupForm.defaultAppImage=Sovelluksen oletuskuvake
 appCenter.adminSetupForm.applications=Sovellukset
-appCenter.adminSetupForm.addNewApp=Lisää sovellus
 appCenter.adminSetupForm.editApp=Muokkaa sovellusta
 appCenter.adminSetupForm.modal.DeleteApp=Oletko varma, että ?
 appCenter.adminSetupForm.createNewApp=Lisää sovellus
