+++ app-center-services/src/main/resources/locale/addon/appcenter_fil.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Mga setting
 appCenter.adminSetupForm.maxFavoriteApps=Bilang ng mga paborito ang pinapayagan
 appCenter.adminSetupForm.defaultAppImage=Default na icon ng application
 appCenter.adminSetupForm.applications=Mga Aplikasyon
-appCenter.adminSetupForm.addNewApp=Magdagdag ng isang application
 appCenter.adminSetupForm.editApp=I-edit ang application
 appCenter.adminSetupForm.modal.DeleteApp=Sigurado ka ba?
 appCenter.adminSetupForm.createNewApp=Magdagdag ng application
