+++ app-center-services/src/main/resources/locale/addon/appcenter_fr.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Paramètres
 appCenter.adminSetupForm.maxFavoriteApps=Nombre de favoris autorisés
 appCenter.adminSetupForm.defaultAppImage=Icône par défaut
 appCenter.adminSetupForm.applications=Applications
-appCenter.adminSetupForm.addNewApp=Ajouter une application
 appCenter.adminSetupForm.editApp=Modifier l'application
 appCenter.adminSetupForm.modal.DeleteApp=Êtes-vous sûr?
 appCenter.adminSetupForm.createNewApp=Ajouter une application
