+++ app-center-services/src/main/resources/locale/addon/appcenter_he.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=הגדרות
 appCenter.adminSetupForm.maxFavoriteApps=מספר המועדפים מותר
 appCenter.adminSetupForm.defaultAppImage=סמל יישום ברירת מחדל
 appCenter.adminSetupForm.applications=יישומים
-appCenter.adminSetupForm.addNewApp=הוסף אפליקציה
 appCenter.adminSetupForm.editApp=ערוך את האפליקציה
 appCenter.adminSetupForm.modal.DeleteApp=אתה בטוח?
 appCenter.adminSetupForm.createNewApp=הוסף אפליקציה
