+++ app-center-services/src/main/resources/locale/addon/appcenter_hr.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=postavke
 appCenter.adminSetupForm.maxFavoriteApps=Broj dozvoljenih favorita
 appCenter.adminSetupForm.defaultAppImage=Zadana ikona aplikacije
 appCenter.adminSetupForm.applications=Prijave
-appCenter.adminSetupForm.addNewApp=Dodajte aplikaciju
 appCenter.adminSetupForm.editApp=Uredi aplikaciju
 appCenter.adminSetupForm.modal.DeleteApp=Jeste li sigurni?
 appCenter.adminSetupForm.createNewApp=Dodaj aplikaciju
