+++ app-center-services/src/main/resources/locale/addon/appcenter_hu.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Beállítások
 appCenter.adminSetupForm.maxFavoriteApps=Kedvencek engedélyezett száma
 appCenter.adminSetupForm.defaultAppImage=Alapértelmezett alkalmazás ikon
 appCenter.adminSetupForm.applications=Alkalmazások
-appCenter.adminSetupForm.addNewApp=Alkalmazás hozzáadása
 appCenter.adminSetupForm.editApp=Alkalmazás szerkesztése
 appCenter.adminSetupForm.modal.DeleteApp=Biztos vagy ebben ?
 appCenter.adminSetupForm.createNewApp=Alkalmazás hozzáadása
