+++ app-center-services/src/main/resources/locale/addon/appcenter_id.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Pengaturan
 appCenter.adminSetupForm.maxFavoriteApps=Jumlah favorit yang diizinkan
 appCenter.adminSetupForm.defaultAppImage=Ikon aplikasi bawaan
 appCenter.adminSetupForm.applications=Aplikasi
-appCenter.adminSetupForm.addNewApp=Tambah aplikasi
 appCenter.adminSetupForm.editApp=Edit aplikasi
 appCenter.adminSetupForm.modal.DeleteApp=Apakah Anda yakin?
 appCenter.adminSetupForm.createNewApp=Tambah Aplikasi
