+++ app-center-services/src/main/resources/locale/addon/appcenter_in.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Settings
 appCenter.adminSetupForm.maxFavoriteApps=Number of favorites allowed
 appCenter.adminSetupForm.defaultAppImage=Default application icon
 appCenter.adminSetupForm.applications=Applications
-appCenter.adminSetupForm.addNewApp=Add an application
 appCenter.adminSetupForm.editApp=Edit application
 appCenter.adminSetupForm.modal.DeleteApp=Are you sure ?
 appCenter.adminSetupForm.createNewApp=Add application
