+++ app-center-services/src/main/resources/locale/addon/appcenter_it.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Impostazioni
 appCenter.adminSetupForm.maxFavoriteApps=Numero di preferiti consentiti
 appCenter.adminSetupForm.defaultAppImage=Icona dell'applicazione predefinita
 appCenter.adminSetupForm.applications=Applicazioni
-appCenter.adminSetupForm.addNewApp=Aggiunga un'applicazione
 appCenter.adminSetupForm.editApp=Modifichi l'applicazione
 appCenter.adminSetupForm.modal.DeleteApp=È sicura ?
 appCenter.adminSetupForm.createNewApp=Aggiunga un'applicazione
