+++ app-center-services/src/main/resources/locale/addon/appcenter_ja.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=設定
 appCenter.adminSetupForm.maxFavoriteApps=お気に入りの数を許可する
 appCenter.adminSetupForm.defaultAppImage=デフォルトのアプリアイコン
 appCenter.adminSetupForm.applications=アプリケーション
-appCenter.adminSetupForm.addNewApp=アプリケーションを追加
 appCenter.adminSetupForm.editApp=アプリケーションを編集
 appCenter.adminSetupForm.modal.DeleteApp=よろしいですか？
 appCenter.adminSetupForm.createNewApp=アプリケーションを追加
