+++ app-center-services/src/main/resources/locale/addon/appcenter_ko.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=설정
 appCenter.adminSetupForm.maxFavoriteApps=허용되는 즐겨찾기 수
 appCenter.adminSetupForm.defaultAppImage=기본 애플리케이션 아이콘
 appCenter.adminSetupForm.applications=응용
-appCenter.adminSetupForm.addNewApp=애플리케이션 추가
 appCenter.adminSetupForm.editApp=애플리케이션 수정
 appCenter.adminSetupForm.modal.DeleteApp=확실합니까?
 appCenter.adminSetupForm.createNewApp=애플리케이션 추가
