+++ app-center-services/src/main/resources/locale/addon/appcenter_lt.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Nustatymai
 appCenter.adminSetupForm.maxFavoriteApps=Leidžiamas mėgstamiausių skaičius
 appCenter.adminSetupForm.defaultAppImage=Numatytoji programos piktograma
 appCenter.adminSetupForm.applications=Programos
-appCenter.adminSetupForm.addNewApp=Pridėti programą
 appCenter.adminSetupForm.editApp=Redaguoti programą
 appCenter.adminSetupForm.modal.DeleteApp=Ar tu tikras?
 appCenter.adminSetupForm.createNewApp=Pridėti programą
