+++ app-center-services/src/main/resources/locale/addon/appcenter_nl.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Instellingen
 appCenter.adminSetupForm.maxFavoriteApps=Aantal toegestane favorieten
 appCenter.adminSetupForm.defaultAppImage=Standaard applicatie icoon
 appCenter.adminSetupForm.applications=Sollicitaties
-appCenter.adminSetupForm.addNewApp=Een applicatie toevoegen
 appCenter.adminSetupForm.editApp=Bewerk applicatie
 appCenter.adminSetupForm.modal.DeleteApp=Weet u het zeker?
 appCenter.adminSetupForm.createNewApp=Applicatie toevoegen
