+++ app-center-services/src/main/resources/locale/addon/appcenter_no.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Innstillinger
 appCenter.adminSetupForm.maxFavoriteApps=Antall favoritter tillatt
 appCenter.adminSetupForm.defaultAppImage=Standard programikon
 appCenter.adminSetupForm.applications=Applikasjoner 
-appCenter.adminSetupForm.addNewApp=Legg til en applikasjon
 appCenter.adminSetupForm.editApp=Rediger applikasjonen
 appCenter.adminSetupForm.modal.DeleteApp=Er du sikker?
 appCenter.adminSetupForm.createNewApp=Legg til applikasjon
