+++ app-center-services/src/main/resources/locale/addon/appcenter_pl.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Ustawienia
 appCenter.adminSetupForm.maxFavoriteApps=Liczba dozwolonych ulubionych
 appCenter.adminSetupForm.defaultAppImage=Domyślna ikona aplikacji
 appCenter.adminSetupForm.applications=Aplikacje
-appCenter.adminSetupForm.addNewApp=Dodaj aplikację
 appCenter.adminSetupForm.editApp=Edytuj aplikację
 appCenter.adminSetupForm.modal.DeleteApp=Czy jesteś pewien?
 appCenter.adminSetupForm.createNewApp=Dodaj aplikację
