+++ app-center-services/src/main/resources/locale/addon/appcenter_pt_BR.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Configurações
 appCenter.adminSetupForm.maxFavoriteApps=Número de favoritos permitidos
 appCenter.adminSetupForm.defaultAppImage=Ícone de aplicação padrão
 appCenter.adminSetupForm.applications=Aplicações
-appCenter.adminSetupForm.addNewApp=Adicionar um Aplicativo
 appCenter.adminSetupForm.editApp=Editar aplicação
 appCenter.adminSetupForm.modal.DeleteApp=Você tem certeza?
 appCenter.adminSetupForm.createNewApp=Adicionar aplicação
