+++ app-center-services/src/main/resources/locale/addon/appcenter_pt_PT.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Configurações
 appCenter.adminSetupForm.maxFavoriteApps=Número de favoritos permitidos
 appCenter.adminSetupForm.defaultAppImage=Ícone padrão do aplicativo
 appCenter.adminSetupForm.applications=Aplicações
-appCenter.adminSetupForm.addNewApp=Adicionar uma aplicação
 appCenter.adminSetupForm.editApp=Editar aplicativo
 appCenter.adminSetupForm.modal.DeleteApp=Tem certeza?
 appCenter.adminSetupForm.createNewApp=Adicionar aplicação
