+++ app-center-services/src/main/resources/locale/addon/appcenter_ro.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Setări
 appCenter.adminSetupForm.maxFavoriteApps=Numărul de favorite permis
 appCenter.adminSetupForm.defaultAppImage=Pictogramă a aplicației implicită
 appCenter.adminSetupForm.applications=Cereri
-appCenter.adminSetupForm.addNewApp=Adăugați o aplicație
 appCenter.adminSetupForm.editApp=Editați aplicația
 appCenter.adminSetupForm.modal.DeleteApp=Ești sigur ?
 appCenter.adminSetupForm.createNewApp=Adaugă aplicație
