+++ app-center-services/src/main/resources/locale/addon/appcenter_ru.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Настройки
 appCenter.adminSetupForm.maxFavoriteApps=Разрешенное количество избранных
 appCenter.adminSetupForm.defaultAppImage=Значок приложения по умолчанию
 appCenter.adminSetupForm.applications=Заявления
-appCenter.adminSetupForm.addNewApp=Добавить приложение
 appCenter.adminSetupForm.editApp=Редактировать приложение
 appCenter.adminSetupForm.modal.DeleteApp=Вы уверены?
 appCenter.adminSetupForm.createNewApp=Добавить приложение
