+++ app-center-services/src/main/resources/locale/addon/appcenter_sk.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=nastavenie
 appCenter.adminSetupForm.maxFavoriteApps=Poznámky
 appCenter.adminSetupForm.defaultAppImage=Ikona predvolenej aplikácie
 appCenter.adminSetupForm.applications=Aplikácie
-appCenter.adminSetupForm.addNewApp=Pridajte aplikáciu
 appCenter.adminSetupForm.editApp=Upraviť aplikáciu
 appCenter.adminSetupForm.modal.DeleteApp=Si si istý?
 appCenter.adminSetupForm.createNewApp=Pridať aplikáciu
