+++ app-center-services/src/main/resources/locale/addon/appcenter_sl.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Nastavitve
 appCenter.adminSetupForm.maxFavoriteApps=Dovoljeno število priljubljenih
 appCenter.adminSetupForm.defaultAppImage=Privzeta ikona aplikacije
 appCenter.adminSetupForm.applications=Aplikacije
-appCenter.adminSetupForm.addNewApp=Dodajte aplikacijo
 appCenter.adminSetupForm.editApp=Uredi aplikacijo
 appCenter.adminSetupForm.modal.DeleteApp=Ste prepričani?
 appCenter.adminSetupForm.createNewApp=Dodaj aplikacijo
