+++ app-center-services/src/main/resources/locale/addon/appcenter_sq.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=crwdns25064:0crwdne25064:0
 appCenter.adminSetupForm.maxFavoriteApps=crwdns25066:0crwdne25066:0
 appCenter.adminSetupForm.defaultAppImage=crwdns25068:0crwdne25068:0
 appCenter.adminSetupForm.applications=crwdns25070:0crwdne25070:0
-appCenter.adminSetupForm.addNewApp=crwdns25072:0crwdne25072:0
 appCenter.adminSetupForm.editApp=crwdns25074:0crwdne25074:0
 appCenter.adminSetupForm.modal.DeleteApp=crwdns25076:0crwdne25076:0
 appCenter.adminSetupForm.createNewApp=crwdns25078:0crwdne25078:0
