+++ app-center-services/src/main/resources/locale/addon/appcenter_sv_SE.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Inställningar
 appCenter.adminSetupForm.maxFavoriteApps=Antal tillåtna favoriter
 appCenter.adminSetupForm.defaultAppImage=Standard applikationsikon
 appCenter.adminSetupForm.applications=Applikationer
-appCenter.adminSetupForm.addNewApp=Lägg till en app
 appCenter.adminSetupForm.editApp=Redigera applikation
 appCenter.adminSetupForm.modal.DeleteApp=Är du säker?
 appCenter.adminSetupForm.createNewApp=Lägg till applikation
