+++ app-center-services/src/main/resources/locale/addon/appcenter_th.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=การตั้งค่า
 appCenter.adminSetupForm.maxFavoriteApps=จำนวนรายการโปรดที่อนุญาต
 appCenter.adminSetupForm.defaultAppImage=ไอคอนแอปพลิเคชันเริ่มต้น
 appCenter.adminSetupForm.applications=การใช้งาน
-appCenter.adminSetupForm.addNewApp=เพิ่มแอปพลิเคชัน
 appCenter.adminSetupForm.editApp=แก้ไขแอปพลิเคชัน
 appCenter.adminSetupForm.modal.DeleteApp=คุณแน่ใจเหรอ?
 appCenter.adminSetupForm.createNewApp=เพิ่มแอปพลิเคชัน
