+++ app-center-services/src/main/resources/locale/addon/appcenter_tr.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Ayarlar
 appCenter.adminSetupForm.maxFavoriteApps=İzin verilen favori sayısı
 appCenter.adminSetupForm.defaultAppImage=Varsayılan uygulama simgesi
 appCenter.adminSetupForm.applications=Uygulamalar
-appCenter.adminSetupForm.addNewApp=Bir uygulama ekleyin
 appCenter.adminSetupForm.editApp=Uygulama düzenle
 appCenter.adminSetupForm.modal.DeleteApp=Emin misiniz?
 appCenter.adminSetupForm.createNewApp=Uygulama ekle
