+++ app-center-services/src/main/resources/locale/addon/appcenter_uk.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Налаштування
 appCenter.adminSetupForm.maxFavoriteApps=Дозволена кількість обраних
 appCenter.adminSetupForm.defaultAppImage=Значок додатку за замовчуванням
 appCenter.adminSetupForm.applications=Додатки
-appCenter.adminSetupForm.addNewApp=Додати додаток
 appCenter.adminSetupForm.editApp=Редагувати додаток
 appCenter.adminSetupForm.modal.DeleteApp=Ви впевнені?
 appCenter.adminSetupForm.createNewApp=Додати додаток
