+++ app-center-services/src/main/resources/locale/addon/appcenter_vi.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=Cài đặt
 appCenter.adminSetupForm.maxFavoriteApps=Số lượng mục yêu thích được phép
 appCenter.adminSetupForm.defaultAppImage=Biểu tượng ứng dụng mặc định
 appCenter.adminSetupForm.applications=Các ứng dụng
-appCenter.adminSetupForm.addNewApp=Thêm một ứng dụng
 appCenter.adminSetupForm.editApp=Chỉnh sửa ứng dụng
 appCenter.adminSetupForm.modal.DeleteApp=Bạn có chắc không?
 appCenter.adminSetupForm.createNewApp=Thêm ứng dụng
