+++ app-center-services/src/main/resources/locale/addon/appcenter_zh_CN.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=设置
 appCenter.adminSetupForm.maxFavoriteApps=允许收藏的数量
 appCenter.adminSetupForm.defaultAppImage=默认应用程序图标
 appCenter.adminSetupForm.applications=应用程序
-appCenter.adminSetupForm.addNewApp=添加应用程序
 appCenter.adminSetupForm.editApp=编辑应用程序
 appCenter.adminSetupForm.modal.DeleteApp=您确定吗？
 appCenter.adminSetupForm.createNewApp=添加应用程序
