+++ app-center-services/src/main/resources/locale/addon/appcenter_zh_TW.properties
@@ -37,7 +37,6 @@ appCenter.adminSetupForm.generalSettings=設定
 appCenter.adminSetupForm.maxFavoriteApps=允許的收藏數
 appCenter.adminSetupForm.defaultAppImage=預設應用程式圖標
 appCenter.adminSetupForm.applications=應用領域
-appCenter.adminSetupForm.addNewApp=新增應用程式
 appCenter.adminSetupForm.editApp=編輯申請
 appCenter.adminSetupForm.modal.DeleteApp=你確定嗎 ？
 appCenter.adminSetupForm.createNewApp=新增應用程式
