+++ app-center-webapps/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -74,6 +74,12 @@
         <module>commonVueComponents</module>
       </depends>
       <depends>
+        <module>applicationToolbarComponent</module>
+      </depends>
+      <depends>
+        <module>appLauncherBundle</module>
+      </depends>
+      <depends>
         <module>adminSetupPortletBundle</module>
       </depends>
     </module>
