+++ app-center-webapps/src/main/webapp/vue-apps/adminSetup/initComponents.js
@@ -14,21 +14,28 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
- 
-import ExoAppCenterModal from './components/modal/ExoAppCenterModal.vue';
-import AdminSetupGeneralParams from './components/AdminSetupGeneralParams.vue';
-import AdminSetupList from './components/AdminSetupList.vue';
-import ApplicationFormDrawer from './components/drawer/ApplicationFormDrawer.vue';
-import ExoSuggester from './components/suggester/ExoSuggester.vue';
+
 import AdminSetupApp from './components/AdminSetup.vue';
 
+import AdminSetupList from './components/main/AdminSetupList.vue';
+import AdminSetupToolbar from './components/main/AdminSetupToolbar.vue';
+import AdminSetupItem from './components/main/AdminSetupItem.vue';
+import AdminSetupMenu from './components/main/AdminSetupMenu.vue';
+
+import ApplicationFormDrawer from './components/drawer/ApplicationFormDrawer.vue';
+import ExoSuggester from './components/suggester/ExoSuggester.vue';
+
+import ExoAppCenterModal from './components/modal/ExoAppCenterModal.vue';
+
 const components = {
-  'adminSetup-generalParams': AdminSetupGeneralParams,
-  'adminSetup-list': AdminSetupList,
+  'app-center-admin-setup': AdminSetupApp,
+  'app-center-admin-toolbar': AdminSetupToolbar,
+  'app-center-admin-item': AdminSetupItem,
+  'app-center-admin-menu': AdminSetupMenu,
+  'app-center-admin-setup-list': AdminSetupList,
   'app-center-modal': ExoAppCenterModal,
   'app-center-form-drawer': ApplicationFormDrawer,
   'app-center-suggester': ExoSuggester,
-  'app-center-admin-setup': AdminSetupApp,
 };
 
 for (const key in components) {
